Don't bundle dpkg or rpm
We require users to provide these tools in $PATH. dpkg and rpm do not provide any portable options, and it would be a lot of hassle to bundle them properly in the AppImage.

The best distributions for building are therefore Debian and its derivatives, which provide packages for both that work pretty well.